Moves Eval to compiler module
- Also, upgrades libraries to achieve scala 2.12 compatibility
- Allows to publish the new compiler module
