+++ .travis.yml
@@ -3,9 +3,7 @@ sudo: required
 services:
 - docker
 scala:
-- 2.10.6
 - 2.11.8
-- 2.12.1
 jdk:
 - oraclejdk8
 cache:
@@ -17,7 +15,7 @@ env:
   global: JAVA_OPTS=-Xmx2g SBT_OPTS="-XX:+UseConcMarkSweepGC -XX:MaxPermSize=512m"
 script:
 - sbt test
-- sbt ++$TRAVIS_SCALA_VERSION 'evaluator-compiler/test'
+- sbt 'project evaluator-compiler' '+test'
 
 before_install:
 - if [ "$TRAVIS_BRANCH" = "master" -a "$TRAVIS_PULL_REQUEST" = "false" ]; then
